fix replace-all, move schema.py into schema/__init__.py
