+++ ludwig/models/trainer.py
@@ -450,11 +450,11 @@ class Trainer(BaseTrainer):
     def clip_grads(self, variables):
         """Applies gradient clipping."""
         if self.gradient_clipping_config.clipglobalnorm:
-            torch.nn.schema_utils.clip_grad_norm_(variables, self.gradient_clipping_config.clipglobalnorm)
+            torch.nn.utils.clip_grad_norm_(variables, self.gradient_clipping_config.clipglobalnorm)
         if self.gradient_clipping_config.clipnorm:
-            torch.nn.schema_utils.clip_grad_norm_(variables, self.gradient_clipping_config.clipglobalnorm)
+            torch.nn.utils.clip_grad_norm_(variables, self.gradient_clipping_config.clipglobalnorm)
         if self.gradient_clipping_config.clipvalue:
-            torch.nn.schema_utils.clip_grad_value_(variables, self.gradient_clipping_config.clipvalue)
+            torch.nn.utils.clip_grad_value_(variables, self.gradient_clipping_config.clipvalue)
 
     def set_base_learning_rate(self, base_learning_rate):
         """Sets the target learning rate, and updates the optimizer learning rate."""
